+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 ### Changes
 - Add default values for `ABCDParams` ([#1](https://github.com/AleksanderWWW/abcd-graph/pull/1))
+
+### Features
+- Add `logger` param to enable logging during graph generation ([#9](https://github.com/AleksanderWWW/abcd-graph/pull/9))